EnergyZero-Terraform/requirements-dev.txt
pytest==8.2.0
pytest-cov==5.0.0
pyspark==3.5.1
delta-spark==3.1.0
